chore: rewrite snippet command to new interface
